feat: add WriteFileWithoutSize method
This commit introduces a new method, `WriteFileWithoutSize`, to the `QRCode` struct. This method allows writing a PNG image of the QR code to a file using a default size, eliminating the need to specify the size explicitly. This provides a convenient way to generate QR code images with a predefined resolution.